all the files are moved
