--- polys_ever/__init__.py
@@ -1,3 +1,0 @@
-
-from polys_ever.poly import otherParaCheck
-from polys_ever.poly import Poly
